requirements.txt
Django==5.2.4
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
